requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.2
Deprecated==1.2.13
Django==4.0.5
django-debug-toolbar==3.4.0
django-templated-mail==1.1.1
factory-boy==3.2.1
h11==0.13.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
kombu==5.2.4
outcome==1.2.0
packaging==21.3
Pillow==9.1.1
pluggy==1.0.0
prompt-toolkit==3.0.29
py==1.11.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
pytest-django==4.5.2
pytest-factoryboy==2.5.0
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tomli==2.0.1
trio==0.21.0
trio-websocket==0.9.2
typing-extensions==4.2.0
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
wsproto==1.1.0
